Moved Stargate.StargateSeedUtility to Utilities.DailySeedUtility maintaing BC.
